Don't show Special:ChallengeUser on Special:SpecialPages for anons, and hide Special:ChallengeView entirely from there
Only registered users may issue challenges. Granted, when an anon accesses Special:ChallengeUser, they are presented two buttons for creating an account or logging in, but still, no point in taunting them.

Special:ChallengeView requires an ID parameter to be passed in order to be useful.
When that is not done, the special page displays a "No challenge specified" message.

Change-Id: I9379d52578a76b2750de0dfd098c587de115ae33