Responsiveness(-ish) tweaks for Special:ChallengeHistory
Both the table and the challenge status filter drop-down menu now fit in the content area, even on skins with a very narrow content area (e.g. Dusk, DuskToDawn).
This change does make the table a bit narrower on MonoBook & friends, but overall that's acceptable and it'll probably need redesigning anyway in the future...

Change-Id: Ia065fbea24ab541c5916c77413360d503f69d89a